*) Code should now compile in single precision mode *) I am rewritting the LDA part. I created a system similar to the exchange in GGA, where one only has to write a formula in terms of rs and zeta. *) Added the second derivatives os a number of LDA functionals.

+++ src/main/one_shot.F90
@@ -53,7 +53,7 @@ contains
     type(hamiltonian_t), target, intent(inout) :: h
 
     integer :: ierr
-    real(8) :: E_tot, E_t, E_ext, E_Hartree, E_x, E_c
+    FLOAT :: E_tot, E_t, E_ext, E_Hartree, E_x, E_c
 
     ! load wave-functions
     call states_allocate_wfns(sys%st, sys%gr%m)
